+++ uno/ingredients/mechanism/GlobalizationMechanism.hpp
@@ -18,8 +18,8 @@ public:
    virtual void initialize(Statistics& statistics, const Problem& problem, Iterate& first_iterate) = 0;
    virtual std::tuple<Iterate, double> compute_acceptable_iterate(Statistics& statistics, const Problem& problem, Iterate& current_iterate) = 0;
 
-   [[nodiscard]] int get_hessian_evaluation_count() const;
-   [[nodiscard]] int get_number_subproblems_solved() const;
+   [[nodiscard]] size_t get_hessian_evaluation_count() const;
+   [[nodiscard]] size_t get_number_subproblems_solved() const;
 
 protected:
    /* references to allow polymorphism */
